updated most int to int64

+++ pkg/quickpay/operation.go
@@ -3,7 +3,7 @@ package quickpay
 import "time"
 
 type Operation struct {
-	ID                   int         `json:"id"`
+	ID                   int64       `json:"id"`
 	Type                 string      `json:"type"`
 	Amount               int64       `json:"amount"`
 	Pending              bool        `json:"pending"`
@@ -15,7 +15,7 @@ type Operation struct {
 	CallbackURL          string      `json:"callback_url"`
 	CallbackSuccess      *bool       `json:"callback_success"`
 	CallbackResponseCode *string     `json:"callback_response_code"`
-	CallbackDuration     *int        `json:"callback_duration"`
+	CallbackDuration     *int64      `json:"callback_duration"`
 	Acquirer             string      `json:"acquirer"`
 	SecureStatus3D       *string     `json:"3d_secure_status"`
 	CallbackAt           *time.Time  `json:"callback_at"`
